+++ src/DearImGui/Raw/Context.hs
@@ -43,5 +43,7 @@ imguiContext = mempty
       , ( TypeName "ImFontGlyphRangesBuilder", [t| ImFontGlyphRangesBuilder |] )
       , ( TypeName "ImGuiListClipper", [t| ImGuiListClipper |] )
       , ( TypeName "ImGuiTableSortSpecs", [t| ImGuiTableSortSpecs |] )
+      , ( TypeName "ImGuiPayload", [t| ImGuiPayload |] )
+      , ( TypeName "ImGuiKeyChord", [t| Int |] )
       ]
   }