+++ ChangeLog.md
@@ -9,6 +9,8 @@
       + Added `ImGui_ImplVulkan_DestroyFontsTexture`, but it shouldn't be needed as it is called from impl internals.
     * Breaking: ImplVulkan moved RenderPass into InitInfo structure.
       + Breaking: Haskell API is now using `Either RenderPass RenderingPipelineCreateInfo` to switch between RP/dynamic rendering.
+- Added lots of missing widgets and their wrappers.
+- Breaking: a few functions in `DearImGui` switched away from using CFloat/CBool wrappers.
 
 ## [2.2.1]